only write passwd file in case run-as is specified

+++ src/main/java/wtf/metio/ilo/os/OSSupport.java
@@ -9,6 +9,7 @@ package wtf.metio.ilo.os;
 
 import wtf.metio.ilo.cli.Executables;
 import wtf.metio.ilo.errors.RuntimeIOException;
+import wtf.metio.ilo.utils.Strings;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -60,16 +61,19 @@ public final class OSSupport {
   }
 
   public static Path passwdFile(final String runAs) {
-    try {
-      final var username = System.getProperty("user.name");
-      final var tempFile = Files.createTempFile("ilo", ".passwd");
-      tempFile.toFile().deleteOnExit();
-      final var content = String.format("%s:x:%s::/home/%s:/bin/bash", username, expand(runAs), username);
-      Files.writeString(tempFile, content);
-      return tempFile.toAbsolutePath();
-    } catch (final IOException exception) {
-      throw new RuntimeIOException(exception);
+    if (Strings.isNotBlank(runAs)) {
+      try {
+        final var username = System.getProperty("user.name");
+        final var tempFile = Files.createTempFile("ilo", ".passwd");
+        tempFile.toFile().deleteOnExit();
+        final var content = String.format("%s:x:%s::/home/%s:/bin/bash", username, expand(runAs), username);
+        Files.writeString(tempFile, content);
+        return tempFile.toAbsolutePath();
+      } catch (final IOException exception) {
+        throw new RuntimeIOException(exception);
+      }
     }
+    return null;
   }
 
   private OSSupport() {
